suitcase-api/pyproject.toml
[project]
name = "suitcase"
version = "0.1.0"
description = "Suitcase"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.10.10",
    "alembic>=1.13.3",
    "apscheduler>=3.10.4",
    "asyncpg>=0.30.0",
    "docx>=0.2.4",
    "fastapi[standard]>=0.115.2",
    "langchain-community>=0.3.2",
    "langchain-groq>=0.2.0",
    "prometheus-client>=0.21.0",
    "pydantic-settings>=2.5.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.0",
    "redis>=5.1.1",
    "requests>=2.32.3",
    "scrapy>=2.11.2",
    "sentence-transformers>=3.2.0",
    "sqlalchemy>=2.0.36",
    "structlog>=24.4.0",
    "supertokens-python>=0.24.3",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "httpx>=0.26.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]
